AutoDJProcessor: Move DeckAttributes into its own file
This is a first step in generalizing the transition logic in
AutoDJProcessor to enable it to also be applied to tracks in
the queue instead of just the current decks.

The code of DeckAttributes has not been changed at all (yet),
and is simply copy-pasted from autodjprocessor.(cpp|h).
