+++ src/library/autodj/track/deckattributes.h
@@ -1,0 +1,132 @@
+#pragma once
+
+#include "audio/frame.h"
+#include "audio/types.h"
+#include "control/controlproxy.h"
+#include "engine/channels/enginechannel.h"
+#include "track/track_decl.h"
+
+class BaseTrackPlayer;
+
+class DeckAttributes : public QObject {
+    Q_OBJECT
+  public:
+    DeckAttributes(int index,
+            BaseTrackPlayer* pPlayer);
+    virtual ~DeckAttributes();
+
+    bool isLeft() const {
+        return m_orientation.get() == static_cast<double>(EngineChannel::LEFT);
+    }
+
+    bool isRight() const {
+        return m_orientation.get() == static_cast<double>(EngineChannel::RIGHT);
+    }
+
+    bool isPlaying() const {
+        return m_play.toBool();
+    }
+
+    void stop() {
+        m_play.set(0.0);
+    }
+
+    void play() {
+        m_play.set(1.0);
+    }
+
+    double playPosition() const {
+        return m_playPos.get();
+    }
+
+    void setPlayPosition(double playpos) {
+        m_playPos.set(playpos);
+    }
+
+    bool isRepeat() const {
+        return m_repeat.toBool();
+    }
+
+    void setRepeat(bool enabled) {
+        m_repeat.set(enabled ? 1.0 : 0.0);
+    }
+
+    mixxx::audio::FramePos introStartPosition() const {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_introStartPos.get());
+    }
+
+    mixxx::audio::FramePos introEndPosition() const {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_introEndPos.get());
+    }
+
+    mixxx::audio::FramePos outroStartPosition() const {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_outroStartPos.get());
+    }
+
+    mixxx::audio::FramePos outroEndPosition() const {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_outroEndPos.get());
+    }
+
+    mixxx::audio::SampleRate sampleRate() const {
+        return mixxx::audio::SampleRate::fromDouble(m_sampleRate.get());
+    }
+
+    mixxx::audio::FramePos trackEndPosition() const {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_trackSamples.get());
+    }
+
+    double rateRatio() const {
+        return m_rateRatio.get();
+    }
+
+    TrackPointer getLoadedTrack() const;
+
+  signals:
+    void playChanged(DeckAttributes* pDeck, bool playing);
+    void playPositionChanged(DeckAttributes* pDeck, double playPosition);
+    void introStartPositionChanged(DeckAttributes* pDeck, double introStartPosition);
+    void introEndPositionChanged(DeckAttributes* pDeck, double introEndPosition);
+    void outroStartPositionChanged(DeckAttributes* pDeck, double outtroStartPosition);
+    void outroEndPositionChanged(DeckAttributes* pDeck, double outroEndPosition);
+    void trackLoaded(DeckAttributes* pDeck, TrackPointer pTrack);
+    void loadingTrack(DeckAttributes* pDeck, TrackPointer pNewTrack, TrackPointer pOldTrack);
+    void playerEmpty(DeckAttributes* pDeck);
+    void rateChanged(DeckAttributes* pDeck);
+    void orientationChanged(DeckAttributes* pDeck);
+
+  private slots:
+    void slotPlayPosChanged(double v);
+    void slotPlayChanged(double v);
+    void slotIntroStartPositionChanged(double v);
+    void slotIntroEndPositionChanged(double v);
+    void slotOutroStartPositionChanged(double v);
+    void slotOutroEndPositionChanged(double v);
+    void slotTrackLoaded(TrackPointer pTrack);
+    void slotLoadingTrack(TrackPointer pNewTrack, TrackPointer pOldTrack);
+    void slotPlayerEmpty();
+    void slotRateChanged(double v);
+    void slotOrientationChanged(double v);
+
+  public:
+    int index;
+    QString group;
+    double startPos;     // Set in toDeck nature
+    double fadeBeginPos; // set in fromDeck nature
+    double fadeEndPos;   // set in fromDeck nature
+    bool isFromDeck;
+    bool loading; // The data is inconsistent during loading a deck
+
+  private:
+    ControlProxy m_orientation;
+    ControlProxy m_playPos;
+    ControlProxy m_play;
+    ControlProxy m_repeat;
+    ControlProxy m_introStartPos;
+    ControlProxy m_introEndPos;
+    ControlProxy m_outroStartPos;
+    ControlProxy m_outroEndPos;
+    ControlProxy m_trackSamples;
+    ControlProxy m_sampleRate;
+    ControlProxy m_rateRatio;
+    BaseTrackPlayer* m_pPlayer;
+};
